Simple test case for server status for the API verified.
