+++ package.json
@@ -4,7 +4,7 @@
   "description": "Mini Digital Bank",
   "main": "./source/server.js",
   "scripts": {
-    "test": "NODE_ENV=test sequelize db:migrate:undo:all &&  NODE_ENV=test sequelize db:migrate &&  NODE_ENV=test sequelize db:seed:all",
+    "test": "NODE_ENV=test sequelize db:migrate:undo:all &&  NODE_ENV=test sequelize db:migrate &&  NODE_ENV=test sequelize db:seed:all && jest",
     "start": "NODE_ENV=test sequelize db:migrate &&  && NODE_ENV=test db:seed:all && node ./source/server.js"
   },
   "repository": {
